Refactor: create a callback format for the visitors. Also removed the dependency (coupling) with the actual nodes, which is a common trick to avoid instanceof (is) and instead do visitor method binding. The problem with that approach is that it leads to coupling between the visitors and the actual data nodes, making it difficult to extend either one without having to do changes to both. Hence the desire for visitors to have a generic accept method which is supposed to handle the entire superset of nodes that may be thrown at them. The risk here, is, of course, that a visitor may receive a node which it does not know how to handle.
In such cases, the visitors must have extension points. And so, here we have defined the open fun acceptUnhandledNode() which is the extension point of the visitors.

I also created a second visitor which directly attaches nodes to the DOM. This is in fact the root use case for the refactoring.